+++ src/love/audio.h
@@ -33,10 +33,6 @@ class audio {
 	SoundData* newSource(const std::string& filename, const std::string& type);
 	SoundData* newSource(const std::string& filename);
 
-	void mixer_render(int16_t *buffer);
-
-	int AUDIO_FRAMES = 735;  // 44100 / 60
-	int WAV_HEADER_SIZE = 44;
 	float m_volume = 1.0f;
 
 	/**