+++ src/love/sound.h
@@ -3,6 +3,7 @@
 
 #include "Types/Audio/SoundData.h"
 #include <vector>
+#include "libretro.h"
 
 namespace love {
 
@@ -12,9 +13,14 @@ namespace love {
 class sound {
 	public:
 	bool load();
+	bool isLoaded();
 	bool hasAudio();
 	void unload();
+	void update();
 	~sound();
+
+
+	// void convert_float_to_s16(int16_t *out, const float *in, size_t samples);
 
 	/**
 	 * Creates a new SoundData.
@@ -26,6 +32,11 @@ class sound {
 	Types::Audio::SoundData* newSoundData(const std::string& filename);
 
 	std::vector<Types::Audio::SoundData*> sounds;
+
+	bool m_loaded = false;
+
+	retro_audio_sample_t audio_cb = NULL;
+	retro_audio_sample_batch_t audio_batch_cb = NULL;
 };
 
 }  // namespace love