+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `love.system.setClipboardText()`
 - Added `love.data.encode()`
 - Added `love.data.decode()`
+- New audio system from libretro-common
 
 ### Chores
 - Added `love.filesystem.mount()` test